Task-13: Created abstract class for tests
