+++ task-13-boyer-moore-algorithm/src/test/java/ru/otus/SubstringSlowSearchTest.java
@@ -1,30 +1,12 @@
 package ru.otus;
 
-import org.junit.jupiter.api.DynamicTest;
-import org.junit.jupiter.api.TestFactory;
 import ru.otus.util.test.TestData;
-import ru.otus.util.test.TestDataSource;
-
-import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class SubstringSlowSearchTest {
+class SubstringSlowSearchTest extends AbstractAlgorithmTest {
 
-	@TestFactory
-	Stream<DynamicTest> find() {
-		return DynamicTest.stream(
-				new TestDataSource("text"),
-				SubstringSlowSearchTest::generateDisplayName,
-				this::executeTest
-		);
-	}
-
-	private static String generateDisplayName(TestData testData) {
-		return "with string length " + testData.getInput().get(1).length();
-	}
-
-	private void executeTest(TestData testData) {
+	protected void executeTest(TestData testData) {
 		final int expected = Integer.parseInt(testData.getOutput().get(0));
 		final int result = SubstringSlowSearch.find(testData.getInput().get(1), testData.getInput().get(0));
